Replace some `Double` with `Long`

+++ src/commonMain/kotlin/entities/UploadCountHistory.kt
@@ -10,4 +10,4 @@ import kotlin.js.JsExport
  * @property uploadCountByDay The list of upload counts per day, over the past 30 days.
  */
 @JsExport @Serializable
-data class UploadCountHistory(val uploadCountByDay: List<Int>? = null)
+data class UploadCountHistory(val uploadCountByDay: List<Long>? = null)
